backend/requirements.txt
Flask==3.1.0
PyJWT==2.12.0
pytest==8.3.5
flake8==7.2.0
